Avalon

+++ Avalon/Avalon.Infra/CharForm.cs
@@ -2,7 +2,7 @@ namespace Avalon.Infra;
 
 public class CharForm : Any
 {
-    public virtual int Execute(int oc)
+    public virtual long Execute(long oc)
     {
         return oc;
     }
